refactor: move standard_board_sizes to untils file
